Update to work with PHP 8.0
The relevant changes to deal with were to namespace and comment parsing.

> T_COMMENT tokens will no longer include a trailing newline. The
> newline will instead be part of a following T_WHITESPACE token. It
> should be noted that T_COMMENT is not always followed by whitespace,
> it may also be followed by T_CLOSE_TAG or end-of-file.
>
> Namespaced names are now represented using the T_NAME_QUALIFIED
> (Foo\Bar), T_NAME_FULLY_QUALIFIED (\Foo\Bar) and T_NAME_RELATIVE
> (namespace\Foo\Bar) tokens. T_NS_SEPARATOR is only used for standalone
> namespace separators, and only syntactially valid in conjunction with
> group use declarations.
- https://www.php.net/manual/en/migration80.incompatible.php#migration80.incompatible.tokenizer
See also https://wiki.php.net/rfc/namespaced_names_as_token

A nuance to the last part about namespaces is that if you parse the
following code with PHP 8 it will return a sequence of T_STRING and
T_NS_SEPARATOR as it did in PHP 7 (presumably because it's not valid in
PHP 8):

namespace RoundTable\
  Knights\
  MontyPython;

I've tried to ensure such namespace declarations can still be parsed in
PHP 8 so that it can be used to help upgrade old code.